21/02/23 Advance Data structures

+++ Java/src/Feb11/Overloading2.java
@@ -18,9 +18,9 @@ public class Overloading2 {
 
     }
     //two float parameters
-     long add(long a, long b)
+     double add(double a, double b)
     {
-        long sum = a+b;
+        double sum = a+b;
         System.out.println("long data type" + sum);
         return sum;
     }
